% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interim_analysis.R
\name{interim_analysis}
\alias{interim_analysis}
\title{Interim analysis with sample size re-estimation}
\usage{
interim_analysis(
  y_trt1,
  y_ct1,
  N_trt,
  N_ct,
  N_ct_max,
  timing_actual,
  alpha0,
  beta0,
  alloc_rate,
  n_endpts,
  SSR_type,
  global_test_type,
  nPM = 100,
  promising_LL = 0.2
)
}
\arguments{
\item{y_trt1}{Stage-1 treatment data (matrix/data.frame: subjects x endpoints).}

\item{y_ct1}{Stage-1 control data (matrix/data.frame: subjects x endpoints).}

\item{N_trt}{Planned total treatment sample size before SSR.}

\item{N_ct}{Planned total control sample size before SSR.}

\item{N_ct_max}{Maximum \emph{re-estimated} control sample size allowed.}

\item{timing_actual}{Interim timing (information fraction, within (0,1))}

\item{alpha0}{One-sided alpha for the power-based SSR formula.}

\item{beta0}{Type II error rate (\eqn{\beta}).}

\item{alloc_rate}{Treatment:control allocation.}

\item{n_endpts}{Number of endpoints.}

\item{SSR_type}{\code{"SSR-Power"}, \code{"SSR-CP"}, or \code{"No SSR"}.}

\item{global_test_type}{\code{"exact OLS"} or \code{"Permutation"}. If
\code{"Permutation"} is used, a random seed is needed to ensure reproducibility.}

\item{nPM}{Number of permutations for the permutation test.}

\item{promising_LL}{Lower bound of the promising zone. Default: 0.2. (Upper bound: 1-beta0)}
}
\value{
An object of class \code{"interim_analysis"}.
}
\description{
Performs an interim global test (exact OLS or permutation). If the interim
statistic crosses the efficacy boundary \code{C1}, the trial stops. Otherwise,
it can do SSR in the promising zone using either \code{SSR-Power} or \code{SSR-CP}.
}
\details{
Stage-1 sample sizes are taken directly from \code{nrow(y_trt1)} and
\code{nrow(y_ct1)}.

This function calls the following \strong{in-package} helpers:
\itemize{
\item \code{get_CP(timing, n, mean_cohen_d, K, sum_rho)}
\item \code{target_CP_get_N_ct(N_ct, timing, mean_cohen_d, K, sum_rho, target_CP)}
\item \code{get_1_PM_mean_z_score(y_mat, trt)}
}
}
\examples{
rho <- matrix(c(
 1.0, 0.1, 0.3, 0.7, 0.1, 0.3,
 0.1, 1.0, 0.7, 0.1, 0.3, 0.7,
 0.3, 0.7, 1.0, 0.1, 0.3, 0.7,
 0.7, 0.1, 0.1, 1.0, 0.1, 0.3,
 0.1, 0.3, 0.3, 0.1, 1.0, 0.7,
 0.3, 0.7, 0.7, 0.3, 0.7, 1.0
 ), nrow = 6, byrow = TRUE)

sim <- simulate_example_one_stage_data(
  n_trt = 40, n_ct = 40,
  n_endpts = 6,
  exp_mean_cohen_d = 0.3,
  mu_ct = c(4, 6, 4, 5, 7, 6),
  sd = c(0.5, 0.5, 1, 1, 2, 2),
  rho = rho,
  seed = 15
)
y_ct1  <- sim$y_ct
y_trt1 <- sim$y_trt

# set.seed(1) # seed needed for "Permutation"
ia <- interim_analysis(
  y_trt1 = y_trt1,
  y_ct1  = y_ct1,
  N_trt = 80,
  N_ct = 80,
  N_ct_max = 160,
  timing_actual = 0.5,
  alpha0 = 0.025,
  beta0 = 0.2,
  alloc_rate = 1,
  n_endpts = 6,
  SSR_type = "SSR-Power",
  global_test_type = "exact OLS"
)
ia
summary(ia)

}
