% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_CP_get_N_ct.R
\name{target_CP_get_N_ct}
\alias{target_CP_get_N_ct}
\title{Root function for CP-based SSR}
\usage{
target_CP_get_N_ct(
  timing,
  n,
  mean_cohen_d,
  K,
  sum_rho,
  target_CP,
  alloc_rate = 1
)
}
\arguments{
\item{timing}{Numeric in (0, 1), interim information fraction.}

\item{n}{Numeric, \strong{candidate} final control group sample size (this is the
variable over which \code{uniroot()} searches).}

\item{mean_cohen_d}{Numeric, observed mean Cohen’s d at interim.}

\item{K}{Integer, number of endpoints.}

\item{sum_rho}{Numeric, sum of all off-diagonal endpoint correlations at
interim, i.e. \eqn{\sum_{p \ne q} \hat\rho_{pq,1}}.}

\item{target_CP}{Numeric in (0, 1), the conditional power you want to hit}

\item{alloc_rate}{Numeric > 0, treatment:control allocation ratio; default \code{1}.}
}
\value{
A single numeric \code{get_CP(...) - target_CP}. Suitable for
\code{uniroot()}.
}
\description{
This is a small helper meant to be used \strong{inside} \code{uniroot()} when
doing sample size re-estimation based on a \emph{target conditional power}.
}
\details{
Given a candidate control sample size \code{n}, it:
\enumerate{
\item computes the conditional power via \code{get_CP(...)} using
Eq. (2.10), and
\item returns \code{CP(n) - target_CP}.
}

Then \code{uniroot()} can be used to find the \code{n} such that
\code{CP(n) - target_CP = 0}.
}
\examples{
\dontrun{
root_res <- stats::uniroot(
  target_CP_get_N_ct,
  lower = 10,
  upper = 1000,
  timing = 0.5,
  mean_cohen_d = 0.3,,
  K = 6,
  sum_rho = 9,
  target_CP = 0.8,
  alloc_rate = 1
  )
n_hat <- ceiling(root_res$root)
}

}
